% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_depths.r
\name{MHD_relative}
\alias{MHD_relative}
\title{Modified Metric Halp-space Depth (Relative Version)}
\usage{
MHD_relative(D, ref_idx)
}
\arguments{
\item{D}{A square distance matrix.}

\item{ref_idx}{Integer vector of reference indices.}
}
\value{
A numeric vector of depth values.
}
\description{
Computes the relative modified Metric Halp-space Depth of observations
with respect to a reference subset.
}
