% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_depths.r
\name{MLD_relative}
\alias{MLD_relative}
\title{Modified Lens Depth (Relative Version)}
\usage{
MLD_relative(D, ref_idx)
}
\arguments{
\item{D}{A square distance matrix.}

\item{ref_idx}{Integer vector of reference indices.}
}
\value{
A numeric vector of depth values.
}
\description{
Computes the relative modified lens depth of observations
with respect to a reference subset.
}
