% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_depths.R
\name{MOD2_relative}
\alias{MOD2_relative}
\title{Modified Metric Oja Depth 2D (Relative Version)}
\usage{
MOD2_relative(D, ref_idx)
}
\arguments{
\item{D}{A square distance matrix.}

\item{ref_idx}{Integer vector of reference indices.}
}
\value{
A numeric vector of depth values.
}
\description{
Computes the relative modified Metric Oja Depth (2D) of observations
with respect to a reference subset.
}
