% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_depths.R
\name{MSD_relative}
\alias{MSD_relative}
\title{Modified Metric Spatial Depth (Relative Version)}
\usage{
MSD_relative(D, ref_idx)
}
\arguments{
\item{D}{A square distance matrix.}

\item{ref_idx}{Integer vector of reference indices.}
}
\value{
A numeric vector of depth values.
}
\description{
Computes the relative modified Metric Spatial Depth of observations
with respect to a reference subset.
}
